+++ internal/gitops/structure/common/types.go
@@ -26,11 +26,9 @@ type StructureConfig struct {
 	Provider             string
 	ClusterBaseTemplates ClusterBaseTemplates
 
-	ClusterBase           string
-	ClusterRelease        string
-	ClusterUserConfig     string
-	DefaultAppsRelease    string
-	DefaultAppsUserConfig string
+	ClusterBase       string
+	Release           string
+	ClusterUserConfig string
 
 	EncryptionKeyPair encryption.KeyPair
 	EncryptionTarget  string
@@ -50,9 +48,6 @@ type StructureConfig struct {
 type ClusterBaseTemplates struct {
 	ClusterAppCr  string
 	ClusterValues string
-
-	DefaultAppsAppCr  string
-	DefaultAppsValues string
 }
 
 type Template struct {