+++ internal/gitops/structure/workload-cluster/templates/template.go
@@ -12,9 +12,6 @@ var appsKustomization string
 //go:embed cluster_userconfig.yaml.tmpl
 var clusterUserConfig string
 
-//go:embed default_apps_userconfig.yaml.tmpl
-var defaultAppsUserConfig string
-
 //go:embed kustomization.yaml.tmpl
 var kustomization string
 
@@ -23,9 +20,6 @@ var patchClusterConfig string
 
 //go:embed patch_cluster_userconfig.yaml.tmpl
 var patchClusterUserconfig string
-
-//go:embed patch_default_apps_userconfig.yaml.tmpl
-var patchDefaultAppsUserconfig string
 
 //go:embed private-key.yaml.tmpl
 var privateKey string
@@ -44,9 +38,7 @@ func GetClusterDirectoryTemplates() []common.Template {
 	return []common.Template{
 		common.Template{Name: "kustomization.yaml", Data: kustomization},
 		common.Template{Name: "cluster_userconfig.yaml", Data: clusterUserConfig},
-		common.Template{Name: "default_apps_userconfig.yaml", Data: defaultAppsUserConfig},
 		common.Template{Name: "patch_cluster_userconfig.yaml", Data: patchClusterUserconfig},
-		common.Template{Name: "patch_default_apps_userconfig.yaml", Data: patchDefaultAppsUserconfig},
 	}
 }